Error handling section : Data annotations on the API object, Fluent validation, Validation middle ware, Axios intercepting errors in response, adding not found, server error pages.

+++ Application/Activities/commands/DeleteActivity.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using Application.Core;
 using Domain;
 using MediatR;
 using Persistence;
@@ -11,21 +12,30 @@ namespace Application.Activities.Commands;
 
 public class DeleteActivity
 {
-	public class Command() : IRequest 
+	public class Command() : IRequest<Result<Unit>>
 	{
 		public required string Id { get; set; }
 	}
 
-    public class Handler(AppDBContext context) : IRequestHandler<Command>
+    public class Handler(AppDBContext context) : IRequestHandler<Command, Result<Unit>>
     {
-        public async Task Handle(Command request, CancellationToken cancellationToken)
+        public async Task<Result<Unit>> Handle(Command request, CancellationToken cancellationToken)
         {
-            var activity = await context.Activities.FindAsync([request.Id],cancellationToken)
-                ??throw new Exception("Activity Not found!!");
+            var activity = await context.Activities.FindAsync([request.Id], cancellationToken);
+
+            if (activity is null)
+            {
+                return Result<Unit>.Failure("Activity not found!", 404);
+            }
 
             context.Activities.Remove(activity);
 
-            await context.SaveChangesAsync(cancellationToken);
+            var result = await context.SaveChangesAsync(cancellationToken) > 0;
+
+            if (!result)
+                return Result<Unit>.Failure("Failed to delete activity", 400);
+
+            return Result<Unit>.Success(Unit.Value);
         }
     }
 }
